requirements.txt
python-dotenv==1.0.0
random-text-generator==1.1.0
requests==2.31.0
django-debug-toolbar==4.2.0
teamapt-monnify==1.1.1
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
idna==3.4
oauthlib==3.2.2
Pillow==9.2.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.5
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
urllib3==1.26.12